+++ NEWS.md
@@ -1,3 +1,5 @@
+# Development
+
 # 3.2.0
 
 ## Breaking changes